Fix fetching wrong folders
