+++ app/gui/vite.config.ts
@@ -57,7 +57,6 @@ export default defineConfig({
       include: fileURLToPath(new URL('./src/dashboard/**/*.tsx', import.meta.url)),
       babel: {
         plugins: [
-          '@babel/plugin-syntax-import-attributes',
           [
             'babel-plugin-react-compiler',
             { target: '18', enablePreserveExistingMemoizationGuarantees: true },
